Added support to train with the whole dataset
Not only with a specified number of atoms, but with all nr of atoms. See prepare_dataset docstring:

load_other_nr_atoms (bool, optional): If set to True, it will load all the data in the parent folder of the specified data folder in `dataset_path` into a unified dataset. Default is False.